+++ packages/expo-router/CHANGELOG.md
@@ -13,6 +13,8 @@
 - Add submenus to link preview context menu ([#37784](https://github.com/expo/expo/pull/37784) by [@Ubax](https://github.com/Ubax))
 - [web] Add styled modals and sheets on web with a custom modal stack using vaul ([#37767](https://github.com/expo/expo/pull/37767) by [@hirbod](https://github.com/hirbod))
 - [web] Add transparent modal support and allow modal stacking ([#37856](https://github.com/expo/expo/pull/37856) by [@hirbod](https://github.com/hirbod))
+- Change modal API to use onRequestClose instead of onClose ([#37762](https://github.com/expo/expo/pull/37762) by [@Ubax](https://github.com/Ubax))
+
 
 ### 🐛 Bug fixes